ath11k: fix multicast rate parameter size
The ATH11K_HW_RATE_CODE() macro uses at least 28 bit to encode the
parameters. An u8 cannot store all these bits.

Fixes: c885aaf227ec ("ath11k: support for multicast rate control")

+++ drivers/net/wireless/ath/ath11k/mac.c
@@ -1577,7 +1577,7 @@ static void ath11k_mac_op_bss_info_changed(struct ieee80211_hw *hw,
 	u16 bitrate;
 	int ret = 0;
 	u8 rateidx;
-	u8 rate;
+	u32 rate;
 
 	mutex_lock(&ar->conf_mutex);
